Let -> Local Definition + Let
Proofs that begin with Let are not to be exported.
Coq 8.5 adds Local Lemma for that, but currently only Local Definition
parses correctly.  This commit turns:

  Let foo : T. Proof. ... Qed.

into

  Local Definition foo : T. Proof. ... Qed. Let foo := foo.

The trailing Let is needed in order to have // pick up
these lemmas.

This change has a factor of 3 on coqc -quick, since Let proofs
cannot be postponed.