workaround a probable bug in ocamlbuild's builtin mldylib rule
ocamlbuild keeps rebuilding .cmxs because a .so is missing, but it
seems the built-in rule is wrong in assuming that a .so will be
produced by the OCaml compiler in the first place (this is
suspiciously similar to upstream bug PR#5100, but I don't yet
understand whether this is the same issue). This patch supersedes the
ocamlbuild built-in rule with a copy of it that does not declare a .so
production, allowing for less recompilation.