Enable parallel build, fix #88
Apparently, the jobserver flags don't show up until we're actually
executing a rule.  So rather than save the old flags, we instead record
whether or not we've seen -B before adding it (actually, whether we've
seen -B or B, since it seems that make stores B rather than -B in
MAKEFLAGS?), and then, later, we filter out B and -B only if we didn't
see them before we added them.