lock set1
The real thing would be to lock group1. The problem is that
hints like `x \in setT` diverge on goals like `1%g \in 1%G`
since the the 1%G of the quotient group is large, and
setT = fun _ => true while (mem 1%G) = fun _ => eq_op x
and Coq tries hard to see if the eq_uo of the quotient group
is the constant function. I could not lock oneg, but set1 was
ok thanks to the inE "abstraction"