[bigop] Add lemma for boolean enumeration in bigops
In some cases, such as in the proof in [1], when we have a bigop
over the boolean enumeration, we need to open the implementation
internals which is not optimal and moreover fragile.

We instead provide a helper lemma that expands a bigop over the
booleans with a concrete enumeration, and which is more
straightforward to use than `big_enumP`.

Note we require commutativity of the operator, this is to safeguard
against changes on the low-level `bool_enumP` enumeration.

[1] https://github.com/math-comp/analysis/blob/7b53e30b53b96e5eb3cabe476b0191353191b52c/theories/altreals/distr.v#L336

Co-authored-by: Amel Kebbouche <akebbouc@outlook.fr>