Add [in A] notation.
Add `[in A]` notation for `fun x => x \in A`, the natural applicative equivalent of a collective predicate `A`.
Replace `(mem A)` and `[mem A]`, which had a similar purpose but require an explicit `/=` or `inE` to turn `mem A x` to `x \in A`, which is sometimes awkward.
Remove the Coq 8.10 compatibility code from `ssrbool.v`, which is actually incompatible with Coq < 8.13.
Add lemmas for `all` and `map`: `allT`, `all_mapT`, `inj_in_map`, `mapK_in`.