Generalize lemmas from monoid to semigroups
Many lemmas on monoids (particularly abelian ones) where already valid
on (abelian) semigroups. This generalization could be improved by
adding semigroups to the monoid hierarchy, when porting to HB.

Some lemmas don't require a monoid with a neutral element but only an
idempotent element. This is particularly useful for min/max that are
idempotent but generally don't have a neutral element.