Add parsing notations for _ = _ :> int and the like
This way `2 = 1 :> int` is parsed as `2%Z = 1%Z` and not `2%R = 1%R`
(that is `Posz 2 = Posz 1` (matching the already existing printing only
rule for `_ = _ :> int`) and not `intmul GRIng.one 2 = GRing.one`).
This should enable to retrieve the parsing behavior of `_ = _ :> int`
before the introduction of the number notation in ring_scope.