Silence warning opaque-let
Introduced in https://github.com/coq/coq/pull/17205
Using `Local Lemma` is not satisfying (exposes the lemma and require
manual introduction of hints). Replacing `Qed` with `Defined` doesn't
seem to incur a noticable slowdown but clutters the proof context with
bodies of the lemmas. So silencing the warning for now until Coq
offers a way to automatically do the `clearbody` at beginning of each
proof.