Adapt to Coq PR #18591: better refolding of addn induces "simpl never" now respected
On subterm "3 + (n - 3)).-2", ".-2" is asking to reduce its argument
twice and the "simpl never" rule does not apply for "+", since "+"
(addn) is in position of argument of a destructor.

We eventually get "1 + (n - 3)" but, without Coq PR #18591, "+" is
actually not refolded so we get instead
"(fix f n := ... body of + ...) 1 (n - 3)", and since "+" is not
refolded, "simpl never" is bypassed, leading to "(n - 3).+1".

With Coq PR #18591, we get "1 + (n - 3)" without "+" being
unfolded. Moreover, "+" is not anymore in position of a destructor, so
the "simpl never" rule applies again, stopping the computation on
"1 + (n - 3)".

It is difficult to evaluate whether there would be a semantics of
"simpl" that would still produce the nicer "(n - 3).+1" in the given
example even though "+" is simpl never. At the current time, I did not
find one not too artificial.